+++ src/FizzWare.NBuilder/Implementation/RangeDeclaration.cs
@@ -7,45 +7,36 @@ namespace FizzWare.NBuilder.Implementation
     [DebuggerDisplay("Range: Start={Start}; End={End}; ({NumberOfAffectedItems} Affected Items)")]
     public class RangeDeclaration<T> : Declaration<T>
     {
-        private readonly int start;
-        private readonly int end;
+        private readonly int _start;
+        private readonly int _end;
 
         public RangeDeclaration(IListBuilderImpl<T> listBuilderImpl, IObjectBuilder<T> objectBuilder, int start, int end)
             : base(listBuilderImpl, objectBuilder)
         {
-            this.start = start;
-            this.end = end;
+            this._start = start;
+            this._end = end;
         }
 
         internal DateTime Created { get; } = DateTime.Now;
 
-        public override int Start
-        {
-            get { return start; }
-        }
+        public override int Start => _start;
 
-        public override int End
-        {
-            get { return end; }
-        }
+        public override int End => _end;
 
         public override void Construct()
         {
-            for (int i = Start; i < End + 1; i++)
+            for (var i = Start; i < End + 1; i++)
                 myList.Add(objectBuilder.Construct(i));
         }
 
         public override void AddToMaster(T[] masterList)
         {
-            for (int i = Start; i < End + 1; i++)
+            for (var i = Start; i < End + 1; i++)
             {
                 AddItemToMaster(myList[i - Start], masterList, i);
             }
         }
 
-        public override int NumberOfAffectedItems
-        {
-            get { return (end + 1) - start; }
-        }
+        public override int NumberOfAffectedItems => (_end + 1) - _start;
     }
 }