+++ tests/FizzWare.NBuilder.Tests/Unit/OperableExtensionTests.cs
@@ -26,6 +26,8 @@ namespace FizzWare.NBuilder.Tests.Unit
             {
             }
 
+            public int Length { get; }
+
             public void CallFunctions(IList<MyClass> masterList)
             {
                 throw new NotImplementedException();